requirement.txt
mysql.connector
flask
uuid
flask_cors